takt: add-e2e-coverage
